pyproject.toml
[project]
name = "climate-change-effect-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "geoalchemy2>=0.18.0",
    "geopandas>=1.1.1",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "scipy>=1.15.2",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
